+++ apps/dine-pleiepenger/src/components/innsendt-søknad-liste/InnsendtSøknadTitle.tsx
@@ -1,9 +1,8 @@
-import { Task } from '@navikt/ds-icons';
 import { BodyShort, HStack, VStack } from '@navikt/ds-react';
-
 import { AppText } from '../../i18n';
 import { InnsendtSøknad } from '../../types/InnsendtSøknad';
 import { formatInnsendtSøknadOpprettetDato } from '../../utils/innsendtSøknadUtils';
+import { TasklistIcon } from '@navikt/aksel-icons';
 
 interface Props {
     søknad: InnsendtSøknad;
@@ -13,7 +12,7 @@ const InnsendtSøknadTitle = ({ søknad }: Props) => {
     return (
         <VStack gap="2">
             <HStack gap="2" align="start" wrap={false}>
-                <Task role="presentation" aria-hidden={true} width="1.25rem" height="1.25rem" />
+                <TasklistIcon role="presentation" aria-hidden={true} width="1.25rem" height="1.25rem" />
                 <div>
                     <BodyShort as="div" size="large" weight="semibold">
                         <AppText id={`sakstype.${søknad.søknadstype}`} />