+++ apps/dine-pleiepenger/src/components/oppdatere-sak-lenker/OppdatereSakLenker.tsx
@@ -1,7 +1,6 @@
-import { PencilWritingIcon } from '@navikt/aksel-icons';
-import { File, Task } from '@navikt/ds-icons';
 import { Heading, HGrid, VStack } from '@navikt/ds-react';
-
+import { PencilWritingIcon, TasklistIcon } from '@navikt/aksel-icons';
+import { FileIcon } from '@navikt/aksel-icons';
 import { useAppIntl } from '../../i18n';
 import { browserEnv } from '../../utils/env';
 import SnarveiLinkCard from '../snarvei-link-card/SnarveiLinkCard';
@@ -35,13 +34,13 @@ const OppdatereSakLenker = () => {
                 <HGrid gap="4" columns={{ sm: 1, md: 2 }}>
                     <SnarveiLinkCard
                         href={browserEnv.NEXT_PUBLIC_SKJEMA_ETTERSENDELSE_URL}
-                        icon={<File role="presentation" aria-hidden={true} width="1.5rem" height="1.5rem" />}
+                        icon={<FileIcon role="presentation" aria-hidden={true} width="1.5rem" height="1.5rem" />}
                         title={text('snarveier.ettersend.tittel')}
                         description={text('snarveier.ettersend.tekst')}
                     />
                     <SnarveiLinkCard
                         href={browserEnv.NEXT_PUBLIC_SKJEMA_PLEIEPENGER_URL}
-                        icon={<Task role="presentation" aria-hidden={true} width="1.5rem" height="1.5rem" />}
+                        icon={<TasklistIcon role="presentation" aria-hidden={true} width="1.5rem" height="1.5rem" />}
                         title={text('snarveier.nySøknad.tittel')}
                         description={text('snarveier.nySøknad.tekst')}
                     />