+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/modules/oppgaver/rapporter-inntekt/i18n/nb.ts
@@ -1,13 +1,26 @@
 export const rapporterInntektMessages_nb = {
     'rapporterInntektOppgavetekst.tittel': 'Hei {deltakerNavn}',
-    'rapporterInntektOppgavetekst.intro.1':
-        'Gi oss beskjed hvis du hadde inntekt og fikk utbetalt lønn fra en arbeidsgiver i {måned}.',
+    'rapporterInntektOppgavetekst.intro.1': 'Gi oss beskjed hvis du hadde inntekt i {måned}.',
     'rapporterInntektOppgavetekst.intro.2':
-        'Inntekt er pengene du får utbetalt fra arbeidsgiveren din. I tillegg til lønn, kan det for eksempel være feriepenger, kvelds- og helgetillegg, honorarer og så videre. Les mer om hva inntekt er i <link>skatteloven § 5.10-15</link>.',
-    'rapporterInntektOppgavetekst.instruks':
-        'Kryss av på “Ja”, skriv inn inntekten din og send inn svaret ditt <strong>senest {frist}</strong>.',
-    'rapporterInntektOppgavetekst.ingenInntekt':
-        'Du kan la være å svare eller krysse av på ”Nei” hvis du ikke hadde inntekt i {måned}.',
+        'Inntekt er lønn, men det kan også være for eksempel etterbetaling, feriepenger, overtid og tillegg for ubekevem arbeidstidstid.',
+    'rapporterInntektOppgave.readMore.tittel': 'Les mer om inntekt',
+    'rapporterInntektOppgave.readMore.tekst.1':
+        'Inntekt er som regel lønnen du får fra en arbeidsgiver, men det kan være mange andre ting også. De vanligste formene for inntekt utenom lønn, er',
+    'rapporterInntektOppgave.readMore.liste.1': 'etterbetaling',
+    'rapporterInntektOppgave.readMore.liste.2': 'feriepenger',
+    'rapporterInntektOppgave.readMore.liste.3': 'overtid',
+    'rapporterInntektOppgave.readMore.liste.4': 'tillegg for kveld, natt, helg og helligdag (ubekevem arbeidstid)',
+    'rapporterInntektOppgave.readMore.liste.5': 'tips',
+    'rapporterInntektOppgave.readMore.liste.6': 'frilansinntekt',
+    'rapporterInntektOppgave.readMore.liste.7': 'inntekt fra aksjeselskap (AS)',
+    'rapporterInntektOppgave.readMore.tekst.2':
+        'Pengene du får fra ungdomsprogramytelsen, skal du ikke gi beskjed om her.',
+    'rapporterInntektOppgave.readMore.tekst.3':
+        '<link>Du kan lese mer om hva som regnes som inntekt i skatteloven §§ 5.10 til 5.15.</link>',
+    'rapporterInntektOppgave.intro.3':
+        'Hvis du hadde inntekt, krysser du av for <strong>Ja og sender inn svaret ditt senest {frist}</strong>.',
+    'rapporterInntektOppgave.intro.4':
+        'Hvis du ikke hadde inntekt, krysser du av på <strong>Nei</strong> eller lar være å svare.',
 
     'rapporterInntektOppgavePart.tittel': 'Inntekt i {månedOgÅr}',
     'rapporterInntektOppgavePart.oppsummering': 'Oppsummering',