+++ apps/ungdomsytelse-deltaker/src/apps/søknad/components/VelkommenMelding.tsx
@@ -1,10 +1,10 @@
+import { ExternalLinkIcon } from '@navikt/aksel-icons';
 import { BodyLong, Box, GuidePanel, Heading, ReadMore, VStack } from '@navikt/ds-react';
 import { dateFormatter } from '@navikt/sif-common-utils';
 import { AppText, useAppIntl } from '@shared/i18n';
 import getLenker from '@shared/utils/lenker';
 
 import BehandlingAvPersonopplysningerContent from './BehandlingAvPersonopplysningerContent';
-import ExternalLink from './external-link/ExternalLink';
 
 interface Props {
     fornavn: string;
@@ -37,7 +37,9 @@ const VelkommenMelding = ({ fornavn, startdato }: Props) => {
                             id="velkommenMelding.søknadBeskrivelse"
                             values={{
                                 Lenke: (children) => (
-                                    <ExternalLink href={getLenker().omUngdomsprogramytelsen}>{children}</ExternalLink>
+                                    <ExternalLinkIcon role="presentation" href={getLenker().omUngdomsprogramytelsen}>
+                                        {children}
+                                    </ExternalLinkIcon>
                                 ),
                             }}
                         />