+++ apps/ungdomsytelse-deltaker/src/apps/søknad/pages/VelkommenPage.tsx
@@ -1,11 +1,10 @@
-import { ArrowRightIcon } from '@navikt/aksel-icons';
+import { ArrowRightIcon, ExternalLinkIcon } from '@navikt/aksel-icons';
 import { BodyLong, Box, Button, Checkbox, CheckboxGroup, VStack } from '@navikt/ds-react';
 import { AppText, useAppIntl } from '@shared/i18n';
 import DefaultPageLayout from '@shared/pages/layout/DefaultPageLayout';
 import getLenker from '@shared/utils/lenker';
 import { useState } from 'react';
 
-import ExternalLink from '../components/external-link/ExternalLink';
 import SøknadHeader from '../components/søknad-header/SøknadHeader';
 import VelkommenMelding from '../components/VelkommenMelding';
 import { useSøknadContext } from '../hooks/context/useSøknadContext';
@@ -43,7 +42,9 @@ const VelkommenPage = () => {
                                 id="velkommenPage.infoStemmer"
                                 values={{
                                     Lenke: (children: string) => (
-                                        <ExternalLink href={getLenker().rettOgPlikt}>{children}</ExternalLink>
+                                        <ExternalLinkIcon role="presentation" href={getLenker().rettOgPlikt}>
+                                            {children}
+                                        </ExternalLinkIcon>
                                     ),
                                 }}
                             />