+++ apps/ungdomsytelse-deltaker/src/apps/søknad/steg/kontonummer/KontonummerSteg.tsx
@@ -1,3 +1,4 @@
+import { ExternalLinkIcon } from '@navikt/aksel-icons';
 import { Alert, BodyLong, BodyShort, Heading, Radio, RadioGroup } from '@navikt/ds-react';
 import { YesOrNo } from '@navikt/sif-common-core-ds/src';
 import { FormLayout } from '@navikt/sif-common-ui';
@@ -5,7 +6,6 @@ import { getYesOrNoValidator } from '@navikt/sif-validation';
 import AriaLiveRegion from '@shared/components/aria-live-region/AriaLiveRegion';
 import { AppText, useAppIntl } from '@shared/i18n';
 import getLenker from '@shared/utils/lenker';
-import ExternalLink from '@søknad/components/external-link/ExternalLink';
 import SøknadSteg from '@søknad/components/søknad-steg/SøknadSteg';
 import SkjemaFooter from '@søknad/components/steg-skjema/SkjemaFooter';
 import { useSøknadContext } from '@søknad/hooks/context/useSøknadContext';
@@ -73,9 +73,11 @@ const KontonummerSteg = () => {
                                                 id="kontonummerSteg.kontonummer.stemmerIkke.info"
                                                 values={{
                                                     Lenke: (children) => (
-                                                        <ExternalLink href={getLenker().personopplysninger}>
+                                                        <ExternalLinkIcon
+                                                            role="presentation"
+                                                            href={getLenker().personopplysninger}>
                                                             {children}
-                                                        </ExternalLink>
+                                                        </ExternalLinkIcon>
                                                     ),
                                                 }}
                                             />
@@ -98,7 +100,9 @@ const KontonummerSteg = () => {
                                     id="kontonummerSteg.harIkkeKontonummer.info.2"
                                     values={{
                                         Lenke: (children) => (
-                                            <ExternalLink href={getLenker().endreKontonummer}>{children}</ExternalLink>
+                                            <ExternalLinkIcon role="presentation" href={getLenker().endreKontonummer}>
+                                                {children}
+                                            </ExternalLinkIcon>
                                         ),
                                     }}
                                 />
@@ -118,7 +122,9 @@ const KontonummerSteg = () => {
                                     id="kontonummerSteg.kontonummerInfoMangler.info.2"
                                     values={{
                                         Lenke: (children) => (
-                                            <ExternalLink href={getLenker().endreKontonummer}>{children}</ExternalLink>
+                                            <ExternalLinkIcon role="presentation" href={getLenker().endreKontonummer}>
+                                                {children}
+                                            </ExternalLinkIcon>
                                         ),
                                     }}
                                 />